Forgot to commit the other file in the last commit.
I don't want to mess around with git at this time of the night.

+++ Assets/Code/Scripts/MainGame/Player/PlayerDeath.cs
@@ -22,7 +22,7 @@ public class PlayerDeath : MonoBehaviour {
 	
 	public void HandleDeath() {
 		
-		GoogleFrontend.Save(Time.timeSinceLevelLoad);
+		GoogleFrontend.Save((int) Time.timeSinceLevelLoad);
 		
 		GameTracker.Active.PutValue("score", Convert.ToString(ScoreDisplayObject.Score));
 		GameTracker.Active.PutValue("time", Convert.ToString(Time.timeSinceLevelLoad));
